Fix mozbrowser race conditions. Fixes #72.
Browser events are currently fired in the global scope and we don't know which frame they are fired for. This passes a callback into each frame for it to call, so it can pass back the frame ID whenever we get a mozbrowser event. Each frame will now be generated with an incrementing ID, and we can filter out updates if they came from an inactive frame.