Initial boilerplate for modal dialogs.
This does not yet actually show anything, but gives us some initial boilerplate that we could use for displaying alert(), prompt() and confirm() calls. I was having a difficult time actually testing this due to content focusing issues. I need to figure out a solution for that, and I think the solution is likely going to be implementing an automated testing framework.